feat: GameResultRecord 클래스명을 Scoreboard로 수정 및 참가자의 수익 구하는 로직 추가
